aオプションとrオプションのテストコードを実装した。テストも通った。
