+++ 08.ls_object/test/test_object_ls.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require 'test/unit'
-require_relative '../object_ls.rb'
+require_relative '../files_collection.rb'
 
 class TEST_OBJECT_LS < Test::Unit::TestCase
 
@@ -17,7 +17,22 @@ class TEST_OBJECT_LS < Test::Unit::TestCase
       
     TEXT
     #ターゲットパスを渡したらそのパスに対してlsコマンドを実行する。空のときはカレントディレクトリでlsコマンドを実行する。
-    files_collection = FilesCollection.new(TARGET_PATHNAME)
+    options = {}
+    files_collection = FilesCollection.new(options, TARGET_PATHNAME)
+    assert_equal expected, files_collection.output_files
+  end
+
+  def test_object_ls_all_reverse
+    expected = <<~TEXT.chomp
+      jjjj  e       .k  
+      i     d       ..  
+      h     c       .   
+      ggg   bbbbbb      
+      f     a           
+      
+    TEXT
+    options = {:all=>true, :reverse=>true}
+    files_collection = FilesCollection.new(options, TARGET_PATHNAME)
     assert_equal expected, files_collection.output_files
   end
 end
